+++ mapsforge-samples-android/src/main/java/org/mapsforge/samples/android/ItemDetailFragment.java
@@ -17,12 +17,10 @@
 package org.mapsforge.samples.android;
 
 import android.os.Bundle;
-import android.os.Environment;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
+import androidx.fragment.app.Fragment;
 import org.mapsforge.map.android.util.AndroidUtil;
 import org.mapsforge.map.android.view.MapView;
 import org.mapsforge.map.layer.cache.TileCache;
@@ -118,7 +116,7 @@ public class ItemDetailFragment extends Fragment {
     }
 
     protected File getMapFileDirectory() {
-        return Environment.getExternalStorageDirectory();
+        return getContext().getExternalFilesDir(null);
     }
 
     protected String getMapFileName() {